biolab1 latest

+++ cosap/pipeline_runner/_pipeline_runner.py
@@ -73,8 +73,9 @@ class PipelineRunner:
 
         config_yaml_path = join_paths(workdir, "config.yaml")
 
-        with open(config_yaml_path, "w") as config_yaml:
-            yaml.dump(config, config_yaml, default_flow_style=False)
+        if not os.path.isfile(config_yaml_path):
+            with open(config_yaml_path, "w") as config_yaml:
+                yaml.dump(config, config_yaml, default_flow_style=False)
 
         snakemake_unlock_dir_command = [
             "snakemake",
